adding optional architecture flag into Ivy files. This seems to work for MathSAT.
This intends to no longer have multiple architectures mixed in one directory in Ivy,
but use a subdirectory "<ARCH>" in the repository directory and also in the cache directory.
The installed library is still located in the same directory as before
(`lib/java/runtime-mathsat/`).
This does not allow to install multiple architectures (`aarch64`, `x86_64`) with identical filenames
(`libmathsat5j.so`).
Installing MathSAT for different OS (with different file endings) are still possible for `x86_64`
(for backward compatibility).

This is not yet tested for using JavaSMT as a library.